+++ src/main/java/com/example/demo/payload/UserRegisterDTO.java
@@ -23,7 +23,7 @@ public class UserRegisterDTO {
     private String addressLine2;
     private String phoneNumber;
     private String postalCode;
-    private Integer localityId;
+    private Long localityId;
     private String locality;
     private String county;
 }
